omit back button on all pages
So far the wizard is not ment to reconfigure RMT, but just for a
first-time setup. Because of that, some pages (database, ssl)
automatically go to the next page when their job is done already when
their are initialized. This broke the back button behavior. Since there
is no good reason to go back anyway at the moment, the fix for now is to
not even show a back button.