Add package version to user agent
To enhance the 'User-Agent' used in the request headers, this commit
adds the module constant `RMT::Version`, which must always match the
package version specified in 'package/yast2-rmt.spec'.

Additionally, a unit test was created to ensure the versions always
match each other; if there is a version bump in packge spec file, the
RSpec tests will break if 'src/lib/rmt.rb' is not updated accordingly.