Configure rules for Rubocop
After removing the gem 'scc-codestyle' (outdated version), Rubocop's
default configuration caused a lot of warnings/errors when linting the
project.

This commit adds rules to 'rubocop.yml' so that the current code passes
the linting process in the CI. Though I agree some of the rules are
important to keep a good style for the project, the idea is just to
adjust our CI process for Tumbleweed at the moment.